fixed crash issue if fragments switched too fast added firebase crashlytics for crash reports

+++ app/src/main/java/com/jjkaps/epantry/ui/Shopping/ShoppingFragment.java
@@ -1,9 +1,7 @@
 package com.jjkaps.epantry.ui.Shopping;
 
-import android.app.Activity;
 import android.app.Dialog;
 import android.os.Bundle;
-import android.provider.DocumentsContract;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.MenuItem;
@@ -12,8 +10,6 @@ import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.Button;
-import android.widget.CheckBox;
-import android.widget.Checkable;
 import android.widget.CheckedTextView;
 import android.widget.EditText;
 import android.widget.ImageButton;
@@ -23,10 +19,7 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentActivity;
-import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
 
 import com.google.android.gms.tasks.OnCompleteListener;
@@ -35,13 +28,9 @@ import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentReference;
-import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.firestore.Query;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.QuerySnapshot;
 import com.jjkaps.epantry.MainActivity;
@@ -52,8 +41,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-
-import io.opencensus.metrics.LongGauge;
 
 public class ShoppingFragment extends Fragment {
     private static final String TAG = "ShoppingFragment";
@@ -75,10 +62,10 @@ public class ShoppingFragment extends Fragment {
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        myDialog = new Dialog(this.getContext());
 
         ShoppingViewModel shoppingViewModel = new ViewModelProvider(this).get(ShoppingViewModel.class);
         final View root = inflater.inflate(R.layout.fragment_shopping, container, false);
+        myDialog = new Dialog(root.getContext());
         if(getActivity() != null && ((MainActivity) getActivity()).getSupportActionBar() !=null){
             View view = Objects.requireNonNull(((MainActivity) getActivity()).getSupportActionBar()).getCustomView();
             TextView name = view.findViewById(R.id.name);
@@ -89,6 +76,10 @@ public class ShoppingFragment extends Fragment {
         btClearAll = root.findViewById(R.id.bt_clearAll);
         imgBtAdd = root.findViewById(R.id.ibt_add);
         txtNullList = root.findViewById(R.id.txt_nullList);
+        final PopupMenu popupMenu = new PopupMenu(getActivity().getApplicationContext(), imgBtAdd);
+        final List<String> shoppingItem = new ArrayList<>();
+        final List<Boolean> itemChecked = new ArrayList<>();
+        final ArrayAdapter arrayAdapter = new ArrayAdapter(root.getContext(), android.R.layout.simple_list_item_multiple_choice,shoppingItem);
 
 
         //Retrieve ShoppingList
@@ -96,17 +87,18 @@ public class ShoppingFragment extends Fragment {
                 .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
                     @Override
                     public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                        List<String> shoppingItem = new ArrayList<>();
-                        List<Boolean> itemChecked = new ArrayList<>();
                         final ListView listView_shopItem = root.findViewById(R.id.listView_shopItem);
                         //Update check status
-                        if (task.isSuccessful() && task.getResult().size() != 0) {
+                        if (task.isSuccessful() && task.getResult() != null && task.getResult().size() != 0) {
                             txtNullList.setVisibility(View.INVISIBLE);
                             for (QueryDocumentSnapshot document : task.getResult()) {
-                                shoppingItem.add(document.get("Name").toString());
+                                Object s = document.get("Name");
+                                if(s != null){
+                                    shoppingItem.add(s.toString());
+                                }
                                 itemChecked.add((Boolean) document.get("Checked"));
                             }
-                            ArrayAdapter arrayAdapter = new ArrayAdapter(getContext(), android.R.layout.simple_list_item_multiple_choice,shoppingItem);
+
                             listView_shopItem.setAdapter(arrayAdapter);
                             listView_shopItem.setChoiceMode(ListView.CHOICE_MODE_MULTIPLE);
                             for(int i=0; i<arrayAdapter.getCount(); i++) {
@@ -121,7 +113,7 @@ public class ShoppingFragment extends Fragment {
                                                 @Override
                                                 public void onComplete(@NonNull Task<QuerySnapshot> task) {
                                                     CheckedTextView v = (CheckedTextView) view;
-                                                    if (task.isSuccessful()) {
+                                                    if (task.isSuccessful() && task.getResult() != null) {
                                                         for (QueryDocumentSnapshot document : task.getResult()) {
                                                             Log.d(TAG, document.getId() + " => " + document.getData());
                                                             shopListRef.document(document.getId()).update("Checked", v.isChecked());
@@ -143,7 +135,6 @@ public class ShoppingFragment extends Fragment {
         imgBtAdd.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(final View view) {
-                PopupMenu popupMenu = new PopupMenu(getActivity().getApplicationContext(), imgBtAdd);
                 popupMenu.getMenuInflater().inflate(R.menu.popup_menu, popupMenu.getMenu());
 
                 popupMenu.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
@@ -172,7 +163,7 @@ public class ShoppingFragment extends Fragment {
                         .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
                             @Override
                             public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                                if (task.isSuccessful()) {
+                                if (task.isSuccessful() && task.getResult() != null) {
                                     for (QueryDocumentSnapshot document : task.getResult()) {
                                         shopListRef.document(document.getId()).delete();
                                     }
@@ -213,13 +204,12 @@ public class ShoppingFragment extends Fragment {
                         .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
                             @Override
                             public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                                if (task.isSuccessful()) {
+                                if (task.isSuccessful() && task.getResult() != null) {
                                     if (task.getResult().size()!=0){
                                         for (QueryDocumentSnapshot document : task.getResult()) {
                                             Toast.makeText(getContext(), item+" Exists!", Toast.LENGTH_SHORT).show();
                                         }
                                         inputItem.setText(null);
-                                        return;
                                     }
                                     //if not exist then add
                                     else {
